修复浮点数bug和统计工作量代码

+++ src/main/java/org/jplus/pojo/yjsjyxm/Yjshxxm.java
@@ -22,7 +22,7 @@ public class Yjshxxm {
     private String xmmc;
 
 //到账经费
-    private Integer dzjf;
+    private Float dzjf;
 
 //项目排名
     private Integer xmpm;
